# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Category {
  id: string;
  name: string;
  slug: string;
  description?: string;
  merchantCount: number;
}

export interface MerchantSummary {
  id: string;
  slug: string;
  name: string;
  category: string;
  tagline?: string;
  rating: number;
  reviewCount: number;
  location: string;
  imageUrl: string;
  featured?: boolean;
  startingPrice?: number;
}

export interface Service {
  id: string;
  merchantSlug: string;
  name: string;
  description: string;
  price: number;
  duration: string;
  imageUrl: string;
  popular?: boolean;
}

export type Merchant = MerchantSummary & ({
  description: string;
  availability: string;
  /** @nullable */
  website?: string | null;
  services: Service[];
});

export interface MerchantInput {
  /** @minLength 2 */
  name: string;
  category: string;
  tagline?: string;
  description: string;
  location: string;
  imageUrl: string;
  availability?: string;
}

export interface ServiceInput {
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  duration: string;
  imageUrl: string;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  confirmed: 'confirmed',
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Order {
  id: string;
  merchantName: string;
  merchantSlug: string;
  serviceName: string;
  customerName: string;
  total: number;
  status: OrderStatus;
  createdAt: string;
  scheduledAt: string;
  paymentMethod?: string;
}

export interface CheckoutInput {
  merchantSlug: string;
  serviceId: string;
  customerName: string;
  customerEmail: string;
  scheduledAt: string;
  notes?: string;
}

export interface CheckoutSession {
  checkoutUrl: string;
  orderId: string;
}

export interface Review {
  id: string;
  customerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  createdAt: string;
}

export interface ReviewInput {
  customerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
}

export type ListMerchantsParams = {
category?: string;
search?: string;
location?: string;
/**
 * @maximum 50
 */
limit?: number;
};

export type ListOrdersParams = {
role?: ListOrdersRole;
status?: string;
};

export type ListOrdersRole = typeof ListOrdersRole[keyof typeof ListOrdersRole];


export const ListOrdersRole = {
  customer: 'customer',
  merchant: 'merchant',
} as const;