# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: marketplace
    description: Marketplace discovery and commerce
  - name: merchants
    description: Merchant storefronts and dashboard
  - name: orders
    description: Checkout and order history
  - name: reviews
    description: Customer reviews
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /categories:
    get:
      operationId: listCategories
      tags: [marketplace]
      summary: List service categories
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
  /merchants:
    get:
      operationId: listMerchants
      tags: [marketplace]
      summary: Discover merchants
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        "200":
          description: Merchants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MerchantSummary"
    post:
      operationId: createMerchant
      tags: [merchants]
      summary: Create a merchant storefront
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MerchantInput"
      responses:
        "201":
          description: Merchant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
  /merchants/{slug}:
    get:
      operationId: getMerchant
      tags: [marketplace, merchants]
      summary: Get a merchant storefront
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Merchant storefront
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Merchant"
        "404":
          description: Merchant not found
  /merchants/{slug}/services:
    get:
      operationId: listMerchantServices
      tags: [marketplace, merchants]
      summary: List services for a merchant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [merchants]
      summary: Add a service to a storefront
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Service created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders for the current account
      parameters:
        - name: role
          in: query
          schema:
            type: string
            enum: [customer, merchant]
            default: customer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
  /orders/checkout:
    post:
      operationId: createCheckout
      tags: [orders]
      summary: Start a secure checkout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "201":
          description: Checkout session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /merchants/{slug}/reviews:
    get:
      operationId: listMerchantReviews
      tags: [reviews]
      summary: List reviews for a merchant
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Leave a review
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Category:
      type: object
      required: [id, name, slug, merchantCount]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        merchantCount: { type: integer }
    MerchantSummary:
      type: object
      required: [id, slug, name, category, rating, reviewCount, location, imageUrl]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        category: { type: string }
        tagline: { type: string }
        rating: { type: number }
        reviewCount: { type: integer }
        location: { type: string }
        imageUrl: { type: string }
        featured: { type: boolean }
        startingPrice: { type: number }
    Merchant:
      allOf:
        - $ref: "#/components/schemas/MerchantSummary"
        - type: object
          required: [description, availability, services]
          properties:
            description: { type: string }
            availability: { type: string }
            website: { type: ["string", "null"] }
            services:
              type: array
              items: { $ref: "#/components/schemas/Service" }
    MerchantInput:
      type: object
      required: [name, category, description, location, imageUrl]
      properties:
        name: { type: string, minLength: 2 }
        category: { type: string }
        tagline: { type: string }
        description: { type: string }
        location: { type: string }
        imageUrl: { type: string }
        availability: { type: string }
    Service:
      type: object
      required: [id, merchantSlug, name, description, price, duration, imageUrl]
      properties:
        id: { type: string }
        merchantSlug: { type: string }
        name: { type: string }
        description: { type: string }
        price: { type: number }
        duration: { type: string }
        imageUrl: { type: string }
        popular: { type: boolean }
    ServiceInput:
      type: object
      required: [name, description, price, duration, imageUrl]
      properties:
        name: { type: string }
        description: { type: string }
        price: { type: number, minimum: 0 }
        duration: { type: string }
        imageUrl: { type: string }
    Order:
      type: object
      required: [id, merchantName, merchantSlug, serviceName, customerName, total, status, createdAt, scheduledAt]
      properties:
        id: { type: string }
        merchantName: { type: string }
        merchantSlug: { type: string }
        serviceName: { type: string }
        customerName: { type: string }
        total: { type: number }
        status: { type: string, enum: [confirmed, pending, completed, cancelled] }
        createdAt: { type: string }
        scheduledAt: { type: string }
        paymentMethod: { type: string }
    CheckoutInput:
      type: object
      required: [merchantSlug, serviceId, customerName, customerEmail, scheduledAt]
      properties:
        merchantSlug: { type: string }
        serviceId: { type: string }
        customerName: { type: string }
        customerEmail: { type: string, format: email }
        scheduledAt: { type: string }
        notes: { type: string }
    CheckoutSession:
      type: object
      required: [checkoutUrl, orderId]
      properties:
        checkoutUrl: { type: string }
        orderId: { type: string }
    Review:
      type: object
      required: [id, customerName, rating, comment, createdAt]
      properties:
        id: { type: string }
        customerName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }
        createdAt: { type: string }
    ReviewInput:
      type: object
      required: [customerName, rating, comment]
      properties:
        customerName: { type: string }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string }