# lib/api-zod/src/generated/types/checkoutInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutInput {
  merchantSlug: string;
  serviceId: string;
  customerName: string;
  customerEmail: string;
  scheduledAt: string;
  notes?: string;
}