# lib/api-zod/src/generated/types/merchant.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { MerchantSummary } from './merchantSummary';
import type { Service } from './service';

export type Merchant = MerchantSummary & ({
  description: string;
  availability: string;
  /** @nullable */
  website?: string | null;
  services: Service[];
});