# lib/api-zod/src/generated/types/merchantInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MerchantInput {
  /** @minLength 2 */
  name: string;
  category: string;
  tagline?: string;
  description: string;
  location: string;
  imageUrl: string;
  availability?: string;
}