# lib/api-zod/src/generated/types/merchantSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface MerchantSummary {
  id: string;
  slug: string;
  name: string;
  category: string;
  tagline?: string;
  rating: number;
  reviewCount: number;
  location: string;
  imageUrl: string;
  featured?: boolean;
  startingPrice?: number;
}