# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: string;
  merchantName: string;
  merchantSlug: string;
  serviceName: string;
  customerName: string;
  total: number;
  status: OrderStatus;
  createdAt: string;
  scheduledAt: string;
  paymentMethod?: string;
}