# lib/api-zod/src/generated/types/review.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Review {
  id: string;
  customerName: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  comment: string;
  createdAt: string;
}