# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: string;
  merchantSlug: string;
  name: string;
  description: string;
  price: number;
  duration: string;
  imageUrl: string;
  popular?: boolean;
}