# lib/api-zod/src/generated/types/serviceInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceInput {
  name: string;
  description: string;
  /** @minimum 0 */
  price: number;
  duration: string;
  imageUrl: string;
}